refactor(auth): delete old auth ui code (#5040)
Anything that is still used was moved to src/login.
