--- packages/core/src/auth/ui/vue/shared.css
@@ -1,31 +1,0 @@
-/* Shared */
-
-button,
-.border-common {
-    border-style: solid;
-    border-width: 0.2rem;
-    border-radius: 1rem;
-    border-color: transparent;
-    font-size: 1rem;
-    padding: 1em;
-}
-
-/*  */
-.container-background {
-    background-color: var(--vscode-sideBar-background);
-}
-
-.main-text-color {
-    color: var(--vscode-foreground);
-}
-
-.text-link-color {
-    color: var(--vscode-textLink-foreground);
-}
-
-body div,
-body h1,
-body h2,
-body h3 {
-    color: var(--vscode-foreground);
-}
