+++ packages/core/src/login/webview/vue/types.ts
@@ -70,3 +70,20 @@ export type AuthEnabledFeatures = 'awsExplorer' | 'codewhisperer' | 'codecatalys
 type Writeable<T> = { -readonly [U in keyof T]: T[U] }
 export type TelemetryMetadata = Partial<Writeable<AuthAddConnection>>
 export type AuthError = { id: string; text: string }
+
+export type ServiceItemId = 'awsExplorer' | 'codewhisperer' | 'codecatalyst'
+export function isServiceItemId(value: unknown): value is ServiceItemId {
+    return (
+        typeof value === 'string' && (value === 'awsExplorer' || value === 'codewhisperer' || value === 'codecatalyst')
+    )
+}
+
+export type AuthFormId =
+    | 'credentials'
+    | 'builderIdCodeWhisperer'
+    | 'builderIdCodeCatalyst'
+    | 'identityCenterCodeWhisperer'
+    | 'identityCenterCodeCatalyst'
+    | 'identityCenterExplorer'
+    | 'aggregateExplorer'
+    | 'unknown'
